backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.50
pydantic==1.10.13
pydantic-settings==2.1.0
python-multipart==0.0.6
pytesseract==0.3.13
Pillow==10.1.0
PyMuPDF==1.23.7
pdf2image==1.17.0
openai==1.6.1
groq==0.4.2
boto3==1.34.11
python-dotenv==1.0.0
reportlab==4.0.8
aiofiles==23.2.1
httpx==0.25.2
